# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Management API",
        "description": "RESTful API for Risk Management System",
        "contact": {
            "email": "admin@riskmanagement.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/risks/{id}": {
            "get": {
                "tags": [
                    "Risks"
                ],
                "summary": "Get a specific risk",
                "operationId": "56dddbceb97d90c1d5c58e4891526bcb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Risk retrieved successfully"
                    },
                    "404": {
                        "description": "Risk not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Risks"
                ],
                "summary": "Update a risk",
                "operationId": "0954cdc72e5078611c3679107e3f8650",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "likelihood": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "impact": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "owner_id": {
                                        "type": "integer"
                                    },
                                    "department": {
                                        "type": "string"
                                    },
                                    "target_closure_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "actual_closure_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Risk updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Risks"
                ],
                "summary": "Delete a risk",
                "operationId": "8c4637b1416c8a9d8ba528e03954ff00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Risk deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}